modules/function_app/function/requirements.txt
logging
os
re
requests
base64
time
azure-functions